refactor: form BatchView element offsets in Int, not size_t
BatchView::operator() cast both indices to size_t before scaling by the
strides, forcing the strided sweep's address arithmetic to 64-bit. The
batch and matrix dimensions all fit in the kernel's Int (the compact
format targets many small matrices), so type the element strides as Int
and compute i*special + p*panel in Int -- no widening of the loop
induction variables, keeping the strided sweep vectorizable. make_view /
make_const_view take Int strides to match, and the four routine drivers
drop their (size_t) element-stride casts. The large per-group base
offset stays size_t and is applied to the pointer before the view is
built. Mark operator() inline.

No behavior change; full MKL suite (11/11) and portable -DCQR_WITH_MKL=OFF
suite (4/4) pass, clang-format clean.

+++ src/cqr_potrf_compact.hpp
@@ -231,8 +231,7 @@ void potrf_compact_general(bool rowmajor, bool upper, Int n, T *ap, Int ldap, In
         else
             /* strided: sweep the reflectorless potf2 with row stride ldap,
              * column stride 1 -- column-major upper and row-major lower. */
-            potrf_compact_group_strided<T, V, Int>(
-                n, make_view<T, V, Int>(a, static_cast<std::size_t>(ldap), 1));
+            potrf_compact_group_strided<T, V, Int>(n, make_view<T, V, Int>(a, ldap, 1));
     }
 }
